+++ quickstart/eabm_package/plot_scatter.py
@@ -1,1 +1,0 @@
-../../tutorials/eabm_package/python/plot_scatter.py