+++ docs/changelog.rst
@@ -8,6 +8,11 @@ Changelog
 0.5.7 (unreleased)
 ******************
 
+New Features
+============
+- Add quantitative regression test option to the tutorial and quickstart post-processing script (:issue:`329`,
+  :merge:`406`). By `Kyle Brindley`_.
+
 Bug fixes
 =========
 - Update the ``plot_scatter.py`` tutorial and quickstart post-processing script to account for the new dimension in
@@ -16,6 +21,8 @@ Bug fixes
 Internal Changes
 ================
 - Add builder action unit tests (:issue:`364`, :merge:`404`). By `Kyle Brindley`_.
+- Change post processing script name in the tutorials and quickstart template files to match broader scope
+  (:issue:`329`, :merge:`406`). By `Kyle Brindley`_.
 
 ******************
 0.5.6 (2022-11-29)