+++ docs/changelog.rst
@@ -14,6 +14,7 @@ New Features
   :merge:`406`). By `Kyle Brindley`_.
 - Manage Cubit environment ``PATH``-like variables from the project configuration file instead of relying on the user
   environment or a project modulefile (:issue:`367`, :merge:`407`). By `Kyle Brindley`_.
+- Update the expected Cubit version from 15.8 to 16.04 (:issue:`367`, :merge:`407`). By `Kyle Brindley`_.
 
 Bug fixes
 =========
@@ -25,6 +26,8 @@ Internal Changes
 - Add builder action unit tests (:issue:`364`, :merge:`404`). By `Kyle Brindley`_.
 - Change post processing script name in the tutorials and quickstart template files to match broader scope
   (:issue:`329`, :merge:`406`). By `Kyle Brindley`_.
+- Remove Matlab and Cubit environment modification from project modulefile (:issue:`367`, :merge:`407`). By `Kyle
+  Brindley`_.
 
 ******************
 0.5.6 (2022-11-29)